fix(dcd): resolve around AddPoapPanel
